+++ include/reactor-uc/clock_synchronization.h
@@ -6,12 +6,6 @@
 #include "reactor-uc/event.h"
 #include "proto/message.pb.h"
 
-#define CLOCK_SYNC_DEFAULT_PERIOD SEC(1)
-#define CLOCK_SYNC_DEFAULT_KP 0.7            // Default value from linuxptp
-#define CLOCK_SYNC_DEFAULT_KI 0.3            // Default value from linuxptp
-#define CLOCK_SYNC_DEFAULT_MAX_ADJ 200000000 // This is the default max-ppb value for linuxptp
-#define CLOCK_SYNC_INITAL_STEP_THRESHOLD MSEC(100)
-
 typedef struct ClockSynchronization ClockSynchronization;
 typedef struct Environment Environment;
