+++ include/reactor-uc/connection.h
@@ -7,13 +7,11 @@
 #include "reactor-uc/reaction.h"
 #include "reactor-uc/reactor.h"
 #include "reactor-uc/trigger.h"
-#include "reactor-uc/platform.h"
 
 typedef struct Connection Connection;
 typedef struct LogicalConnection LogicalConnection;
 typedef struct PhysicalConnection PhysicalConnection;
 typedef struct DelayedConnection DelayedConnection;
-typedef struct EnclavedConnection EnclavedConnection;
 typedef struct Port Port;
 typedef struct Output Output;
 
@@ -51,21 +49,5 @@ struct DelayedConnection {
 void DelayedConnection_ctor(DelayedConnection *self, Reactor *parent, Port **downstreams, size_t num_downstreams,
                             interval_t delay, ConnectionType type, size_t payload_size, void *payload_buf,
                             bool *payload_used_buf, size_t payload_buf_capacity);
-
-struct EnclavedConnection {
-  Connection super;
-  interval_t delay;
-  ConnectionType type;
-  EventPayloadPool payload_pool;
-  void *staged_payload_ptr;
-  MUTEX_T mutex; // Used to protect the _last_known_tag variable.
-  tag_t _last_known_tag;
-  void (*set_last_known_tag)(EnclavedConnection *self, tag_t tag);
-  tag_t (*get_last_known_tag)(EnclavedConnection *self);
-};
-
-void EnclavedConnection_ctor(EnclavedConnection *self, Reactor *parent, Port **downstreams, size_t num_downstreams,
-                             interval_t delay, ConnectionType type, size_t payload_size, void *payload_buf,
-                             bool *payload_used_buf, size_t payload_buf_capacity);
 
 #endif
