+++ lfc/core/src/main/kotlin/org/lflang/generator/uc/UcClockSyncGenerator.kt
@@ -15,6 +15,7 @@ data class UcClockSyncParameters(
     val Ki: Double = UcClockSyncParameters.DEFAULT_KI,
 ) {
   companion object {
+    // Default values for clock sync params. Taken from linuxptp.
     const val DEFAULT_DISABLED = false
     const val DEFAULT_GRANDMASTER = false
     const val DEFAULT_PERIOD = 1000000000L
@@ -41,8 +42,14 @@ class UcClockSyncGenerator(
     private val targetConfig: TargetConfig
 ) {
 
+  // Number of neighbors is the same as the number of federated connection bundles.
   private val numNeighbors = connectionGenerator.getNumFederatedConnectionBundles()
+  // We allocate three event per neighbor and an additional two events for the bookkeeping timer.
   val numSystemEvents = numNeighbors * 3 + 2
+
+  // Clock sync is enabled by default. But can be disabled by a global target property (clock-sync:
+  // off) or by
+  // federate-specific annotation.
   val enabled =
       !federate.clockSyncParams.disabled &&
           targetConfig.getOrDefault(ClockSyncModeProperty.INSTANCE) !=
